codes/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==24.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cosine_annealing_warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup.git@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.1
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.13
fastjsonschema==2.21.1
filelock==3.16.1
flash==1.0.3
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.69.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.12.0
jsonargparse==4.36.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
lightning-bolts==0.7.0
lightning-flash==0.8.2
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyDeprecate==0.3.2
pydicom==2.4.4
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==1.9.5
pytorch-lightning-bolts==0.3.2.post1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
sacremoses==0.1.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
segmentation_models_pytorch==0.4.0
Send2Trash==1.8.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
stanza==1.10.1
stringzilla==3.12.5
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
timm==1.0.13
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchmetrics==0.10.3
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.17.0
triton==3.1.0
types-python-dateutil==2.9.0.20250516
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
wandb==0.19.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yarl==1.18.3
zipp==3.21.0
